Split ingress and loadbalancer playbooks
created ingress playbook in playbooks dir.

Note: Will move all playbooks to playbook dir in the future.

Created metalb role to house the metalb tasks. Here some
defaults where set but can be overriden in inventory for
the cluster master.

Ingresses defined now as multiple hosts in the ingress group.
This is where you can defined the variable overrides for the
particular ingress.